YUK-21837 Custom Attributes Data Export Fixed Attribute

+++ yukon-client/yukon-web/src/com/cannontech/web/admin/dao/impl/CustomAttributeDaoImpl.java
@@ -1,11 +1,13 @@
 package com.cannontech.web.admin.dao.impl;
 
+import java.util.ArrayList;
 import java.util.List;
 
 import org.apache.commons.collections4.CollectionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import com.cannontech.common.exception.DataDependencyException;
+import com.cannontech.common.i18n.MessageSourceAccessor;
 import com.cannontech.common.model.Direction;
 import com.cannontech.common.pao.PaoType;
 import com.cannontech.common.pao.attribute.dao.AttributeDao;
@@ -15,17 +17,24 @@ import com.cannontech.common.pao.attribute.model.AttributeAssignment;
 import com.cannontech.common.pao.attribute.model.CustomAttribute;
 import com.cannontech.common.util.SqlStatementBuilder;
 import com.cannontech.database.SqlParameterSink;
+import com.cannontech.database.TypeRowMapper;
 import com.cannontech.database.YukonJdbcTemplate;
 import com.cannontech.database.incrementer.NextValueHelper;
 import com.cannontech.user.YukonUserContext;
 import com.cannontech.web.admin.dao.CustomAttributeDao;
+import com.cannontech.i18n.YukonUserContextMessageSourceResolver;
 
 public class CustomAttributeDaoImpl implements CustomAttributeDao {
     
     @Autowired private YukonJdbcTemplate jdbcTemplate;
     @Autowired private AttributeDao attributeDao;
     @Autowired private NextValueHelper nextValueHelper;
-
+    @Autowired private YukonUserContextMessageSourceResolver messageSourceResolver;
+    private static String deletionFailed = "yukon.web.modules.adminSetup.config.attributes.deletion";
+    private static String deletionFailedDataExport = "yukon.web.modules.adminSetup.config.attributes.deletion.export";
+    //Device data monitor
+    private static String deletionFailedDDM = "yukon.web.modules.adminSetup.config.attributes.deletion.ddm";
+    
     @Override
     public List<AttributeAssignment> getCustomAttributeDetails(List<Integer> attributeIds, List<PaoType> deviceTypes,
             SortBy sortBy, Direction direction) {
@@ -73,6 +82,7 @@ public class CustomAttributeDaoImpl implements CustomAttributeDao {
         SqlParameterSink params = updateSql.update("AttributeAssignment");
         addAssignmentParameters(params, assignment);
         updateSql.append("WHERE AttributeAssignmentId").eq(assignment.getAttributeAssignmentId());
+        jdbcTemplate.update(updateSql);
         attributeDao.cacheAttributes();
         return attributeDao.getAssignmentById(assignment.getAttributeAssignmentId());
     }
@@ -102,15 +112,28 @@ public class CustomAttributeDaoImpl implements CustomAttributeDao {
 
     @Override
     public void deleteCustomAttribute(int attributeId) throws DataDependencyException {
-        SqlStatementBuilder sql = new SqlStatementBuilder();
-        sql.append("DELETE FROM CustomAttribute");
-        sql.append("WHERE AttributeId").eq(attributeId);
-        /* if(true) {
-         *      throw new DataDependencyException("--------DataDependencyException-------");
-         * }
-         */
-        jdbcTemplate.update(sql);
-        attributeDao.cacheAttributes();
+        
+        MessageSourceAccessor messageSourceAccessor = messageSourceResolver.getMessageSourceAccessor(YukonUserContext.system);
+        
+        //Return message: Attribute {0} cannot be deleted, it is currently being used by: Device Data Monitor(s) alpha, beta. Data Export(s) myExportName
+        List<String> dataExportsUsingTheAttribute = new ArrayList<>();
+        SqlStatementBuilder sqlExport = new SqlStatementBuilder();
+        sqlExport.append("SELECT FormatName FROM ArchiveValuesExportFormat format");
+        sqlExport.append("JOIN ArchiveValuesExportAttribute att on att.FormatId = format.FormatId");
+        sqlExport.append("WHERE att.AttributeName").eq(attributeId);
+        dataExportsUsingTheAttribute.addAll(jdbcTemplate.query(sqlExport, TypeRowMapper.STRING));
+        
+        if(dataExportsUsingTheAttribute.isEmpty()) {
+            SqlStatementBuilder sql = new SqlStatementBuilder();
+            sql.append("DELETE FROM CustomAttribute");
+            sql.append("WHERE AttributeId").eq(attributeId);
+            jdbcTemplate.update(sql);
+            attributeDao.cacheAttributes();
+        }
+        
+        if(true) {
+            throw new DataDependencyException("--------DataDependencyException-------");
+        }
     }
 
     @Override
